+++ apps/client/src/app/pages/map/hooks/useHandleLogin.tsx
@@ -5,6 +5,7 @@ function handleLogin(email: string, password: string) {
     email: emailRef.value,
     password: passwordRef.value
   };
+  console.log(userData);
   return userData;
 }
