+++ apps/client/src/app/pages/map/shared-ui-components/Textfile.tsx
@@ -1,20 +1,23 @@
 function TextField({
-  className,
+  classNameDiv,
+  classNameInput,
   type,
   required,
   onInput,
   placeholder,
-  id
+  id,
+  ref
 }: TextFieldProps) {
   return (
-    <div className={className}>
+    <div className={classNameDiv}>
       <input
         type={type}
         required={required}
-        className={className}
+        className={classNameInput}
         onInput={onInput}
         placeholder={placeholder}
         id={id}
+        ref={ref}
       />
     </div>
   );
